+++ src-docs/src/views/accessibility/props.tsx
@@ -1,6 +1,0 @@
-import React, { FunctionComponent } from 'react';
-import { EuiScreenReaderOnlyProps } from '../../../../src/components/accessibility/screen_reader';
-
-export const ScreenReaderOnlyDocsComponent: FunctionComponent<EuiScreenReaderOnlyProps> = () => (
-  <div />
-);