+++ src-docs/src/views/code/code_block_pre.js
@@ -13,9 +13,7 @@ export default () => (
       whiteSpace="pre"
       isCopyable>
       {`export default () => (
-  <div>
-    In this example, the whiteSpace property is set to pre. All the whitespaces will be kept as is and the text only wraps when line breaks are in the content.
-  </div>
+  <div>In this example, the whiteSpace property is set to pre. All the whitespaces will be kept as is and the text only wraps when line breaks are in the content.</div>
 );`}
     </EuiCodeBlock>
   </div>