+++ src-docs/src/views/highlight_and_mark/highlight_and_mark_example.js
@@ -6,6 +6,8 @@ import { GuideSectionTypes } from '../../components';
 
 import { EuiCode, EuiHighlight, EuiMark } from '../../../../src/components';
 
+import { highlightConfig, markConfig } from './playground';
+
 import { Highlight } from './highlight';
 import { Mark } from './mark';
 
@@ -69,4 +71,5 @@ export const HighlightAndMarkExample = {
       demo: <Mark />,
     },
   ],
+  playground: [highlightConfig, markConfig],
 };